experiments with calling openssl from an installed library

+++ callsha/callsha.h
@@ -2,4 +2,4 @@ int SHA256_sodium(unsigned char *in, unsigned char *out, unsigned long long inle
 
 int SHA256_VST(unsigned char *in, unsigned char *out, unsigned long long inlen);
 
-int SHA256_NSS(unsigned char *in, unsigned char *out, unsigned long long inlen);
+int SHA256_NSS(unsigned char *in, unsigned char *out, unsigned long inlen);
